test(plugin-rapl): Updates to move tests files
