+++ plugins/rapl/src/tests/mocks.rs
