+++ plugins/rapl/src/tests/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(test)]
+pub mod mocks;
+
+#[cfg(test)]
+pub mod tests;
